Add comment stripping and whitespace handling to get_file.
This stops interactive mode exiting prematurely when you hit enter on an empty
line. Whitespace before commands is now stripped by the scanf.
Lines whose first non-whitespace character is # are ignored as comments.

This works for file processing as well as interactive as both are handled
by get_file on a FILE * (file or stdin).